requirements.txt
Flask==3.0.3
gunicorn==21.2.0
gspread==6.1.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.32.5
